+++ crates/gen/src/parser/type_tree.rs
@@ -22,7 +22,7 @@ impl TypeEntry {
         match &self.def {
             ElementType::TypeDef(def) => def.clone().with_generics().gen(gen, self.include),
             ElementType::MethodDef(def) => def.gen(gen),
-            ElementType::Field(def) => def.gen(gen),
+            ElementType::Field(def) => gen_field(def, gen),
             _ => unimplemented!(),
         }
     }
