+++ crates/gen/src/tables/field.rs
@@ -47,10 +47,6 @@ impl Field {
         Row::new(row, TableIndex::TypeDef, self.0.file).into()
     }
 
-    pub fn gen(&self, gen: &Gen) -> TokenStream {
-        types::Constant::gen(self, gen)
-    }
-
     pub fn dependencies(&self, include: TypeInclude) -> Vec<TypeEntry> {
         self.signature().kind.definition(include)
     }
