+++ crates/gen/src/types/mod.rs
@@ -5,7 +5,6 @@ mod bstr;
 mod callback;
 mod class;
 mod com_interface;
-mod constant;
 mod delegate;
 mod r#enum;
 mod function;
@@ -27,7 +26,6 @@ pub use bstr::*;
 pub use callback::*;
 pub use class::*;
 pub use com_interface::*;
-pub use constant::*;
 pub use delegate::*;
 pub use function::*;
 pub use handle::*;
