+++ crates/winmd_reader/src/lib.rs
@@ -1,0 +1,7 @@
+pub use std::collections::{BTreeMap, BTreeSet};
+pub use std::iter::FromIterator;
+
+mod parser;
+pub mod tables;
+
+pub use parser::*;
